refactor(3d): split the gaze/playback JS out of flightmap3d_html.py (#383)
The 723-line Camera's Gaze + playback + click-a-spot block moves
verbatim to flightmap3d_gaze_js.py, spliced over a __GAZE_JS__
placeholder exactly where it used to live -- load-time execution order
is unchanged, and app_js reaches the template as a .format() value, so
no brace doubling is involved (mirrors __SHARED_JS__).

Two fold-ins named in the issue ride along, and they are the ONLY
behavioural deltas (verified by diffing the emitted HTML against the
pre-split build):

- One source for the rings: renderGaze called gazeRing directly while
  gazeRingsFor memoised the identical values. gazeRingAt now memoises
  per index -- playback warms only what it draws, the click lookup's
  full scan fills the rest -- and the memo gained `reason` so the
  render path could join it.

- The emitted JS is ASCII-only again: the literal multiply sign and em
  dash (plus two non-ASCII comments) now use \uXXXX escapes or ASCII,
  and a test pins the rule for both modules.
